refactor: TimeTableCell 테스트 -> Timetable 테스트
